webapps/requirements-dev.txt
coderedcms==0.19.0

black==19.10b0
flake8==3.8.3
gunicorn==20.0.4
whitenoise==5.1.0
coverage==5.1
django-environ==0.4.5

# Heroku
# django-toolbelt==0.0.1
# psycopg2-binary==2.8.5
# dj-database-url==0.5.0
# South==1.0.2

# search
# wagtail-whoosh==0.2.4

# other
phonenumbers==8.12.8

# To use with a MySQL database.
# Don't forget specify appropriate version.
# mysqlclient==2.0.1

# PENGGANTI django-compressor==2.4
# pip install rcssmin --install-option="--without-c-extensions"
# pip install rjsmin --install-option="--without-c-extensions"
# pip install django-compressor --upgrade